requirements.txt
Flask==2.0.2
gunicorn
pandas
python-dotenv
